bug 1603308: fix date range for ES lookup
Previously, verifyprocessed would do an ES search using the default date
range which is a week ago to today. If the crash it was looking at was
outside of that range, it wouldn't get found.

This fixes it to look at the day the crash id says the collector picked
it up.