bug 1603308: fix cronrun to save place after every run_time
If cronrun is working on a backfill job that has a bunch of run_times
to get through, then it'll go through all those run_times and if
everything is successful, it'll update the "last_success" field for that
job.

However, since the run_times for a backfilled job start with the
"last_success" datetime, if a job fails half-way through, it'll start
all over again at the beginning. Ugh.

This fixes it so that cronrun saves its place after every successful
run_time. If it fails (ran out of time, some random problem, whatever),
it'll start the run_times at the last successful run_time. This will
drastically improve the situation where a backfill job has to catch up
from many failures.