bug 1597730: remove quit check
We used to do deploys by stopping the processor service, updating the
code, and then starting it again. That was a long time ago. Now we
create a bunch of new processor nodes and then kill off the old
processor nodes. Because of that, we no longer need the quit check code
that allows the processor to shut down nicely.