bug 1552976: fix correlations tab
When the report view page is loaded, the correlations tab is hidden.
JavaScript in the page goes to fetch correlation data and if it gets
some, then it unhides the tab so as to show correlations data. The
problem here is that unhiding the tab causes the tabs to the right to
jump.

This changes the code so that the tab is always shown and when there's
no correlations data, it states there's no correlations data. This fixes
the UI jumping which keeps biting me. It also reduces confusion around
why the tab isn't showing up in some cases.

While doing this, I also unhid the Telemetry Environment tab. That will
now show all the time and when there's no TelemetryEnvironment data,
it'll state there's no data.