bug 1543767: fix topcrashers summary line
The topcrashers summary line was confusing because the math was being
done against a "total_number_of_crashes" that was not the total number
of crashes but instead the user-provided number of signatures to look
at. This fixes everything by pulling the total number from the search
results which is the total number of crash reports that match the given
filters.