bug 1605429: fix stackwalker to work with ModuleSignatureInfo as object
Recently, the crash reporter was changed so it sends a JSON blob with
the annotations in it. When that happened, the type of the
ModuleSignatureInfo value changed from a JSON-encoded string to an
object. stackwalker expected a JSON-encoded string, so if it was an
object, it'd ignore it. Thus it didn't capture the signed-by information
for modules.

This fixes that by supporting both JSON-encoded string and Json::Value
values of ModuleSignatureInfo.