bug 1608111: add single JSON-encoded value bits to specification
Originally, I was thinking the JSON-encoded value bits constitute a new
version of the specification. However, since we decided that the
collector would accept crash reports and the processor would fix them, I
don't think it's necessary to have a new version of the specification.

Thus, this updates the specification with the new bits plus it adds a
very minimal "debugging" section.