bug 1616837: add RustMozCrash to irrelevant list
This adds RustMozCrash and mozglue_static::panic_hook to the irrelevant
list. If they show up in stacks, they'll get skipped over in signature
generation.

This also adds core::ops::function::Fn::call<T> to the prefix list so
it's included in the signature, but generation continues past it.

These changes let us differentiate between causes for crash reports
that are all currently getting a signature of RustMozCrash.