bug 1616194: use json_dump.lsb_release.description for platform pretty version
For crash reports from Linux, the platform pretty version has been
"Linux" which is pretty unhelpful.

This changes it so that it uses the value in
json_dump.lsb_release.description which comes from /etc/lsb_release
information in Linux. When this is available, it's a far more useful
string like "Ubuntu 18.04 LTS".