bug 1619646: add handling for UnicodeEncodeError in truncating
The Elasticsearch crash storage code tries to be really forgiving for
bad data. One of the things it does is truncate absurdly long string
values. Strings are unicode, but are stored in Elasticsearch as utf-8,
so the truncation code encodes as utf-8 to get the length.

This handles the situation where the string value can't be encoded as
utf-8. In that case, probably best to consider it bad dta and replace it
with "BAD DATA"