bug 1608116: add DeNoneRule
This adds a DeNoneRule that will remove keys from the raw crash that
have a None value. These come from crash annotations the crash reporter
packs into a JSON-encoded value in the crash report. Previously, I think
these would end up with empty strings.

The field we seem to be having the most problems with is BuildID.

This also adds some metrics code for the DeNoneRule and DeNullRule so we
can see how often they're kicking off so as to identify bugs in the
crash reporter.