bug 1622932: fix featured versions
The featured versions list is calculated for most products based on
existing crash reports. This allows the list to be dynamic and change
quickly which is helpful during releases. However, this means that
calculating the list is vulnerable to junk data and bugs.

This adjusts the calculation slightly in that it adds a minimum count
threshold. I picked 50 after looking at all the existing products on
Socorro. Going lower increases the likelihood that junk data affects the
featured versions. Going higher increases the likelihood we hit an issue
with less-used products where versions take a really long time to show
up.

We can tweak the number a little as circumstances come up, but I think
if products have a problem with it, they should do what Fennec is doing
and manually maintain a list of featured versions.

Note that the VERSIONS_COUNT_THRESHOLD is set to 0 for the local dev
environment.