bug 1619638: remove java_stack_trace_full before indexing in Elasticsearch
Elasticsearch has problems with the java_stack_trace_full field. We
removed it a while back, but crash reports processed before we removed
it still have the field in the processed crash file.

This adds it to the forbidden_keys list for Elasticsearch. We can remove
the forbidden key in August 2020.