bug 1628311: remove logging line
This removes a logging line so you can redirect the output to
run_mdsw.sh to a file and then use jq and be able to read it.
Previously, you'd have to edit the file to remove the line that isn't
JSON. That makes it harder to use in scripts.