bug 1641316: fix bugzilla bug creation for java errors
If the crash report represents a java error, then there's a
JavaStackTrace annotation and Crash Stats puts that in the crash report.
However, the ``` were in the wrong place in respects to if/for blocks
and so it was missing the end set.

While fixing that, I tweaked the description a bit.