bug 1611131: improve language around protected data and access
Over the years, Crash Stats has evolved how it talks about data that can
only be seen by people who have been vetted and granted access to it.

This goes through and changes all the language to talk about "protected
data" and "protected data access". This updates the url where the
protected data access policy exists and provides a redirect for the
original policy url. This reduces wasted space on the report view and
adds a statement at the top explaining what can be seen and why with
links to documentation on requesting protected data access. This adds
exclamation marks to some sensitive fields--we should add more as we
find them.

The goals of this work are three-fold:

1. use consistent language
2. make it clearer what is and isn't protected
3. make it clearer what users can see depending on whether they're
   logged in or not and whether they've been granted access to
   protected data