bug 1497353: remove "dump" field
The "dump" field was migrated over when we convered super search fields
from being listed in Elasticsearch to being listed in
super_search_fields.py.

I don't know what it was used for, but it doesn't appear to have any
data in it and it's not used in the processor.