bug 1624345: stop saving random data to elasticsearch crashstorage
Previously, we would put everything in the raw and processed crash in
the document and index that in Elasticsearch--even things that were not
specified in super_search_fields. This caused multiple problems with
strings that were invalid keys. We worked through those. This also
caused the problem where we have a bunch of unknown user-provided data
in Elasticsearch.

The one thing where this helps is we can sometimes retroactively make
a field searchable and have some data in Elasticsearch already. This
doesn't happen often enough to be compelling.

This changes the Elasticsearch crash storage such that it removes
anything in the raw and processed crash that isn't specified in
super_search_fields. This moves us another step towards a world where we
have a well defined schema of what we're storing.