bug 1558916: fix issues with featured versions calculation
This fixes featured versions calculation in regards to how versions get
sorted and how "featured_versions" works in the products files.

Now featured versions are inserted to the beginning of the versions list
and will always show up first in the list. This means featured versions
are now displayed in the order specified in the product file.