bug 1659795: add fission indicators to signature/topcrashers reports
This adds indicators to the signature report summary and topcrashers
report to indicate if all of the crash reports have DOMFissionEnabled as
well as an indicator if some of the crash reports have
DOMFissionEnabled.

This also removes the remnants of font-awesome stuff we had before which
wasn't in use and replaces it with an updated font-awesome. font-awesome
is used for the fission icon.