bug 1661605: move minidump-stackwalk to separate repository
We moved minidump-stackwalk code and build tools to
https://github.com/mozilla-services/minidump-stackwalk .

When that repo is tagged, it generates a Docker image and posts it to
https://hub.docker.com/r/mozilla/socorro-minidump-stackwalk/tags .

The Socorro build process now pulls the binaries from that image rather
than build stackwalker and friends every time during the build process.

This simplifies building Socorro, building minidump-stackwalk,
maintaining stackwalker, and debugging all the bits.