bug 1643339: don't search indices older than retention policy
This fixes supersearch so that regardless of the specified date range,
it'll drop the indices that are older than the retention policy.

These indices shouldn't exist, so supersearch would have searched them,
failed, then spent an inordinate amount of time trying to fix the query
by dropping bad indices one at a time.