bug 1667741: add Windows heap failure error handling to irrelevant list
From Gabriele: Crash signatures that include heap failures on Windows
tend to be very, very long as the first few frames are taken by Windows
own error-handling machinery. It might be better to ignore those frames
entirely in signature generation to make the interesting part of the
stack stand out.