bug 1669163: fix formatting for reason/moz_crash_reason in bug links
MozCrashReason only gets generated in cases where the reason is SIGSEGV
or EXCEPTION_BREAKPOINT so the reason isn't helpful. This changes the
template to show MozCrashReason or Reason or nothing.

Bugzilla comments use Markdown for formatting, so we need to put the
MozCrashReason/Reason in ``` so the values aren't kicking off Markdown
formatting inadvertently.