bug 1669247: conform to Dockerflow standards
This changes DOCKER_USERNAME to DOCKER_USER and DOCKER_PASSWORD to
DOCKER_PASS. This switches to using DOCKERHUB_REPO.

This cleans up the "push to dockerhub" step so it's more like the
Dockerflow one. That pushes a "latest" *only* when the branch is main,
pushes a tag *only* when it's a real tag. This stops pushing individual
commit images by sha--those are now "latest" and there's only one of
them.