bug 1676654: prettier, eslint, and get rid of vestiges of GA
Socorro used to use Google Analytics and had a shim around in case
Google Analytics wasn't set up. We removed Google Analytics ages ago,
but I missed the shim. This removes the shim.

I updated eslint and prettier, but didn't lint the code. This fixes
formatting issues with the code as well as some new linting issues that
came up.