bug 1684164: add application_build_id to search list
Fenix puts the GeckoView build id in BuildID and the Fenix build id in
ApplicationBuildID in crash reports. The GeckoView BuildID is
well-formed--it's a long in the form of YYYYMMDDHHMMSS. The Fenix
ApplicationBuildID is a monotonically increasing number, but isn't
structured on a datestamp. We can treat it like a build id and let
people search with it as if it's a long.

This adds it as a public field that's searchable.

This also adds it to the Details and Metadata tabs of the report view.