bug 1674791: update to Python 3.9.1
In the previous attempt to upgrade from Python 3.8 to 3.9, I hit
problems with building lxml and sasl. Since then, the lxml problem went
away.

sasl is a dependency of pyhs2. That library is both unused (we don't use
Hive) and unmaintained (since 2016). I removed that which removed the
sasl and thrift dependencies.

That unblocked upgrading to Python 3.9.