bug 1671276: process breadcrumbs and show on details page
The crash reporter in android-components which Fenix uses captures
breadcrumbs and puts it in the Breadcrumbs annotation.

This adds a processor rule to validate it and add it to the processed
crash and then display it in the Details of the report view.

Note that we tried to use the same structure that Sentry uses, however
Sentry's breadcrumbs structure looks like this:

{
  "breadcrumbs": {
    "values": [
      crumb,
      crumb,
      crumb,
    ]
  }
}

and ours look like this:

{
  "breadcrumbs": [
    crumb,
    crumb,
    crumb,
  ]
}

This code normalizes the Sentry form to ours because the validation is
easier that way. In the future, we should make sure that's what we want
to do and adjust accordingly.

This marks it as protected data, so if you don't have access to that,
you can't see it.

This doesn't make it searchable. I decided not to deal with that until I
know more about how people want to search it (or even if people do want
to search it).