bug 1643339: re-fix cutoff policy code
When someone does a supersearch and specifies a date range, we don't
want to try to search indices that are outside the retention policy. In
September, I wrote some code that I thought correctly removed indices
that weren't valid, but I didn't think about how our index templates
include a week number, but not a year.

This fixes that code so it's not longer doing bad comparisons with index
names and instead figures out the set of valid indices and intersects
that with the set getting searched.