Fix flakey ES test with more state-checking
This one test is failing sometimes. I'm pretty sure it's a
race-condition between the state of ES indices and what the test is
doing to setup the indices it needs for the test.

This adds some health_check calls which should enforce the cluster is
good for the test to continue.