bug 1686103: remove fields not in mapping for index
When saving crash reports, remove fields that aren't in super search
fields or aren't in the mapping of the index the crash report is going
into.

This fixes the issues around adding new fields to super search fields.
Going forward, there won't be any existing data in that field until the
new index is created. We won't have to deal with existing data that was
indexed with the wrong storage_mapping anymore which breaks aggregation
and searching.