bug 1682931: reduce JavaStackTrace confusion
If you don't have protected data access, then on the Details tab, you
see the "JavaStackTrace (redacted)" key with the java_stack_trace value.
It doesn't show up at all on the Metadata tab.

If you do have protected data access, then on the Details tab, you see
JavaStackTrace (Raw) key with the java_stack_trace_raw value and on the
Metadata tab, you see JavaStackTrace with the java_stack_trace_raw
value.

If you have protected data access, this is confusing because you're
seeing both the raw value in the Details tab and the sanitized value in
the Metadata tab and they have the same value which makes it seem like
Crash Stats is not sanitizing the value correctly.

This fixes that by changing the "Metadata" tab to "Crash annotations",
splitting that data into "public" and "protected" sections, and adding
the "protected data" label to the protected data section.

Further, on the Details tab, I added the "protected data" label to
all the items that are protected data including the "Java Stack Trace
(raw)" field if it's showing.

I also improved the description for "Java Stack Trace (raw)" so it
notes that it's protected data.

I think this clarifies what value you're seeing and what you can do with
it. Further, I think this makes the "Metadata" tab a lot clearer along
with a link to the CrashAnnotations.yaml document that describes the
fields there.