bug 1687987: stop using removed fields
I did a pass on the collector and removed fields it was adding to the
raw crash that we didn't need. This implements the other side of that
and removes those fields from super search and the processed crash.

This removes type_tag, legacy_processing, throttle_rate, and timestamp.

This also reworks the DatesAndTimesRule processor rule which used
timestamp to now use submitted_timestamp. I also had to fix the tests
because they're using fictitious data which doesn't make sense and when
I removed timestamp, that broke the tests.

I also switched from using
socorro.lib.datetimeutil.datetime_from_isodate_string to using
datetime.datetime.fromisoformat which is new in Python 3.7 and is the
inverse of datetime.datetime.isoformat() which we use to generate these
values.