bug 1675560: add java exception to report view
This adds the code to show the java_exeption data on the Details tab of
the report view.

It shows the raw version if you have protected data access or if this
is your crash (which I think is impossible since it keys off of the
email address and Fenix doesn't ask for that).

It shows the redacted version in all other cases. The difference between
the two is that the redacted version doesn't show the exception message.