bug 1691393: drop defaults for http requests
This drops the backoff_factor to 0.1 and the default timeout to 3. This
makes it less likely that trying to request a resource from a
non-responsive service (*cough*Bugzilla*cough*) doesn't end up exceeding
Gunicorn's timeout value at which point Gunicorn kills the process.

I also dropped the waitlist generator from `[2, 2, 2, 2, 2]` to `[1, 1,
1, 1, 1]` which might alleviate a similar issue in the crashstorage code
where it's trying to push something to AWS S3 and has to retry a few
times.