bug 1166058: implement update_graphics_pci
This removes the graphics device admin page that lets us upload and
parse the `pci.ids` file That requires the max body upload size to be
higher than the file size and it's not. Rather than fiddle with that, I
redid everything to run as a cron job.

In doing that, I discovered we were parsing the file wrong, so I fixed
that, too.

I updated the graphicsdevice data flow docs.