bug 1696910: fix data links and redo "raw data and minidumps tab"
This redoes the "Raw data and minidumps" tab so that the links are at
the top and a syntax-highlighted and more usable version of the
minidump-stackwalk output is below it.

This fixes the data links so they're generated in the view code using
API endpoints rather than an odd crashstats view. This fixes the
RawCrash model to have a class method for figuring out the binary
filename rather than a static template.

This redoes how we do list-style CSS by adding a couple of
tailwind-inspired classes and putting them in the typography base file.