bug 1696910: reduce data links
There are two permissions involved in showing the links. Originally, I
split the links into two separate sections--one for each permission.
However, that adds a lot of visual junk to the top of the screen and
it's hard to parse.

This reduces the two sections into a single section in a kind of
"whatever--close enough" mode. In reality, there's only one group we use
and that group has both permissions, so while they're two separate
things, it's not the case that someone would have one but not the other.