bug 1672406: add ModuleURLRewriteRule to generalize sym urls
minidump-stackwalk downloads SYM files to symbolicate and unwind the
stack in the minidump it's working on. It caches the SYM files on disk
so as to reduce network traffic.

If the minidump comes from a try build, the SYM file is cached, but if
minidump-stackwalk processes another crash report from the same try
build and uses the cached SYM files, it puts the wrong url in the
"symbol_url" field for the module.

This fixes that by changing all the symbol_url fields to be
https://symbols.mozilla.org/ redirector urls. The redirector will check
all the AWS S3 buckets it knows about and return the url of the one that
contains the SYM file. This fixes this issue as well as reduces some
migration headaches we'll have if we ever move the SYM files in the
future.