Fixed test that verified crash size capture
This is a test that just started failing in CircleCI. I looked into it
and I'm puzzled as to how it's been passing at all.

The raw_crash in the test has two keys but neither of them should be in
a mapping, so they should get removed before the size of the raw crash
is captured. Given that, it should always have been 2 characters--not
55.

I'm out of time into looking at it, so we'll have to leave it as a
mystery for now.