bug 1701357: change default process_type to parent
Previously, we set the process_type value to None if there was no
ProcessType annotation. This is the case for "main" aka "browser" aka
"parent" process crashes. Then in Supersearch we handled "browser"
search as "does not have a ProcessType". That's awkward when you want to
look at an aggregate of the process_type field since "does not exist"
won't show up in the aggregate. Yuck.

This redoes all that nonsense. Now the default process_type is "parent"
because that's the word that's used in the gecko code so we maintain
terminology. All crash reports will have a process_type in the processed
crash going forward. This lets us remove the hard-coded "browser means
does not exist", it fixes the weirdness with aggregates, the Process
Type field shows up in all crash reports in the Details tab, and we're
one step closer to CONQUERING THE TRI-STATE AREA!

The next step, of course, is to remove the "browser" stuff, but I don't
want to do that for a few months, so I'll put that in a bug to do in
July or August or something.