bug 1701357: redo display of process_type in Details tab
There are several values for process_type that a crash report can have.

If it is "content", then we want to additionally show the RemoteType
value, but only if the user has protected data access.

If it is "plugin", then we want to additionally show data about the
plugin process.

This cleans the template code up so that's clearer both in the template
code and also in the rendered template.