bug 1704529: clean up field value truncation code
This fixes the case where async_shutdown_timeout values weren't getting
truncated.

Before indexing data into Elasticsearch, the crashstorage code truncates
the values for "string" and "keyword" fields. The way it figured out
which fields were which was messy, missed a bunch of fields, and didn't
cache intermediary work which would always produce the same results.

This refactors that code up so it's a little better and fixes the
immediate issue in bug #1704529. The get_fields_by_item() function isn't
great, but it's ok for now.