bug 1699492: fix mutation issues in signature generation
Signature generation used to mutate the stack frames adding a
"normalized" key to them. Not only do we not use that, but there's
another bit of code elsewhere that used to remove it because it was
unhelpful.

This gets rid of the normalized caching altogether. It adds a test to
make sure we're not mutating in the future. It reduces some code around
generating signatures from a stack which suggested that these tools had
the same API but was a lie.