bug 1706076: fix elasticsearch valid keys
This removes some levels of caching that aren't helpful anymore now that
get_keys() is cached.

This also fixes get_keys_for_fields() by changing the name to make it
clearer that it's only getting indexable fields and by changing the list
comprehension to filter out fields that have no storage_mapping. Fields
that don't have a storage_mapping don't get added to the mapping and
thus shouldn't be indexed.

After doing that, we had to update a bunch of tests that relied on
fields being indexed that shouldn't be indexed.