bug 1706076: remove dead supersearch fields
We had a bunch of fields defined in super_search_fields.py that came
over from when it was in the DB. These fields had is_exposed and
is_returned set to False, so they're not used in SuperSearch and they're
not returned by SuperSearch, RawCrash, or ProcessedCrash APIs.

I removed them, fixed some tests accordingly, and also added some
additional things to assert for fields and fixed the tests that failed
those new assertions.

I also improved the docs so it's clearer how to define new fields in the
future.