bug 1706076: fix all_valid_fields to work with nested fields
There are nested data structures in the processed crash. Previously, the
code compared a key for a nested item with the top-level part of a key
from the processed_crash which would never match, so we'd lose a bunch
of stuff. This was obviously broken when I fixed some tests so they were
more explicit in what they were testing.

This fixes it so it builds a new raw/processed crash tree using keys
that specify possibly nested elements by using glom which treats the
keys as paths into the data structure.