bug 973894: fix super search so all fields can use exists filter
This fixes super search so all fields can use "exist/does not exist"
filter. Previously, only string fields could.

This makes it a lot easier to search for crash reports that have or
don't have a specific field. Previously, you'd have to hack it in some
cases and in others, it wasn't possible.