bug 1705469: add support for multiple rulesets in a processing pipeline
This adds basic support for multiple rulesets in a processing pipeline
and the ability to specify which ruleset to use when reprocessing a
crash report.

It adds a "regenerate_signature" ruleset which only runs the
SignatureGenerationRule.

It adds a "--ruleset" argument to the reprocess command.

It adds a "ruleset" tag to the process_crash timing metric so we can
differentiate between timings for different processing pipeline
rulesets.

The Reprocessing API validates ruleset names using a Django setting.
Invalid rulset names result in an HTTP 400.

If the processor gets a ruleset name that doesn't match a ruleset, then
it'll log an error, drop the crash id, and move on to the next one in
the queue.

This also renames "raw_dumps" to "dumps" in the processor. It's a
shorter name and we don't have "processed dumps", so we don't need to
specify whether they're raw or not.