Fix docs requirements and remove howto
ReadTheDocs is currently failing to build Socorro docs because it's
having difficulties figuring out requirements to install. Switching to a
requirements.in file and compiling the requirements will probably help
with that.

This also removes the howto document. It had a single runbook in it and
I moved that to the cloudops-docs repo.