bug 1546800: fix supersearch status to show dates for indices
The template for index names is socorro%Y%W. The %W is the week number.
It's hard for me to figure out what the date is for a week, so I tweaked
the code to figure out the start date for that week and show that in the
indices list, too.