bug 1709024: clarify Installations count in Product table of signature report
The signature report has a Product table that shows the count of crash
reports for a specific version and how many installations. The count is
a precise count of a value that has a small domain of options. The
number of installations is a cardinality aggregation of a datestamp
which has a very large domain of options and thus Elasticsearch computes
an approximate. It's possible for the number of installations to exceed
the number of crash reports which makes no sense and is confusing.

This adds "(approx)" to the column header label. It also caps the number
of installations displayed so it's never more than the number of crash
reports.