Add links to integrating crash reporting with product; fix version notes
This adds some links on how to integrate crash reporting into a product.
I don't think people are generally looking for that kind of
documentation here, but it's a natural thing to cover in Socorro's
product-support documentation.

This updates the docs on version requirements and adds some details on
the versioning schemes used by various products currently supported.