bug 1710854: rework cpu_arch calculation
Previously, the cpu_arch field in the processed crash was based solely
on the output from minidump-stackwalk. Further, if there was no output,
then cpu_arch was set to "". The empty string is a problematic value
since it's difficult to search with.

This changes that code in two ways:

1. it changes the default value for cpu_arch and cpu_info to "unknown"
   which is an easier value to work with
2. if there's no minidump-stackwalk output (no minidumps or a minidump
   that can't be parsed), then it looks at Android_CPU_ABI which is an
   annotation that exists for Android crash reports and covers the case
   where it's a Java crash

We can add additional ways to determine the cpu_arch going forward.