bug 1032227: rework sorting for address field
This fixes the address field so that it sorts correctly client-side and
server-side.

It pulls in a minidump-stackwalk that left-pads the address data
according to the minidump width for the architecture of the process.
Once the addresses are left-padded, they sort alphanumerically
correctly. Given that, this removes the pre-existing hexdigit sorting
(which converted from a hex string to an int and sorted on that).

While doing that, I hit some bugs with sorting other things like
graphics adapter values, so I fixed those, too, by generalizing the
sorting fix. This should fix sorting for all the tables in the signature
report for the address, graphcs device, and adapter fields.