bug 1716742: mark last_error_value ERROR_COMMITMENT_LEVEL as OOM
This marks crash reports from WER that don't have a OOMAllocationSize
annotation, but do have a last_error_value as ERROR_COMMITMENT_LEVEL as
OOM crashes.

To do that, I needed access to the crashing thread data. There was a
bunch of code around accessing that which was different, so I
centralized that and added tests.

While doing that, I reworked SignatureRunWatchDog rule so it didn't use
subclassing but was instead its own rule.