bug 1688908: remove your_crash feature
The your_crash feature allowed users who were logged into Crash Stats
(which is email-address based) to view protected data in crash reports
that had an Email value the same as what they were logged in as.

Since we no longer collect Email address data, this no longer works.
Given that, we can remove it. This commit removes that code.