bug #1703407: remove handling for "browser" as a process type
A long time ago, we used "browser" as the default and did some
shenanigans to make it work as such. We transitioned to using "parent"
as the default value and added a processor rule to fill in the default
process type if there wasn't one already. Now all crash reports
have a process type and we no longer use "browser".

This removes the special casing for handling "browser" as a process type
and covering crash reports that didn't have one in the crash
annotations.