bug 1726032: add ADR log to docs
This adds an ADR log to the docs. It's a table that lists summary
information derived from the .md file and then links to the .md file on
GitHub.

I did it this way because then we don't have to have a markdown parser.
If we ever switch to MyST, we should rethink this.

This also adds a CSS override for the width of the content section on
the HTML page. sphinx_rtd_theme restricts it to 800px and that's really
frustrating because some things are wider than that.