bug 1732414: extract distributionId
This pulls out the TelemetryEnvironment.partner.distributionId value and
puts it in the processed_crash as distribution_id. This value represents
which vendor build the product.

If it's not there at all (no TelemetryEnvironment, etc), we don't know
and mark "unknown".

If it's null, we mark "mozilla".

We use the value as is for everything else.

This allows us to search and aggregate on distribution_id.