bug 1737643: build rust-minidump minidump-stackwalker
We're about to go into a period of time where we're migrating from the
minidump-stackwalk built in C++ using Breakpad to the rust-minidump
minidump-stackwalk built in Rust.

This adds building the rust-minidump minidump-stackwalk as an additional
stage in building Socorro. The way it's done should take advantage of
layer caching, so it's not getting built redundantly.

It puts the resulting binary minidump-stackwalk in /stackwalk-rust/. I
did that for parity with the existing stackwalker. We may move it later
on down the road.

Note that while this does build rust-minidump minidump-stackwalk and
puts it in /stackwalk-rust/, it doesn't do anything with it.