bug 1737878: return normalized frames in signature result
This tweaks signature generation such that if the CSignatureRule
generates normalized versions of the frames in order to build the
signature, it returns those frames in `result.extra["normalized_frames"]`.

Notes for future me: the normalized frames were already in
`result.extra["proto_signature"]` as a pipe-delimited mess. If someone
needed individual frames, splitting that string isn't safe since frames
can have pipes in them because everything is awesome.

While I was doing that, I also fixed the verbose output and
markdown-formatted output of the signature command.