bug 1740397: rework CrashingThreadInfoRule
The CrashingThreadRule set crashedThread in the processed crash, but
that's not used anywhere. We do have a bunch of places that need the
crashing thread index and they do slightly different calculations to get
it.

I changed this rule to set `processed_crash["crashing_thread"]` which
has the index of the crashing thread or None.

I changed some other rules to use this information.