bug 1737691: add new MinidumpStackwalkRule
This adds a new MinidumpStackwalkRule that runs the rust-minidump
minidump-stackwalk binary on minidumps to walk the stack, symbolicate
symbols, and extract other information.

This adds a configuration option `processor.minidump_stackwalk` for
flipping between BreakpadStackwalkerRule2015 and MinidumpStackwalkRule.
If it's set to "rust", then it uses MinidumpStackwalkRule. All other
values including being unset will use BreakpadStackwalkerRule2015.

This also fixes a bunch of code looking at the json_dump output to
better handle when a key exists, but the value is None.