Improve uptime_ts and uptime field docs
This improves the description of the uptime_ts field by copying the
description for the crash annotation it's based on from
CrashAnnotations.yaml.

This also improves the description for the uptime field by adding some
details for how the value is calculated.

This isn't great. We really need a way to annotate fields and expose
those annotations such that we can capture historical knowledge for
these fields, bugs related to them, etc. But this is good enough for
now.