bug 1737691: skip processing for 0-byte dump files
This tweaks the code to skip minidump processing for 0-byte dump files.
There's no reason to do all the work if there's nothing there to
process.

This also improves the processor notes for downstream rules that look
for data generated by minidump processing. If the data ain't there, then
there's nothing for those rules to do. I reduced the language, added
prefixes, and in some cases, nixed the notes altogether.