bug 1737691: remove cruft from docker image
The Dockerfile copies /tmp/set_up_ubuntu/sh into the image and then runs
it. It should remove it as well. This fixes that.

The Dockerfile created directories /tmp/symbols/cache and
/tmp/symbols/tmp but those are only used in the local dev environment.
It's better if the MinidumpStackwalkRule creates the directories it
needs for symbols cache and symbols tmp. That way we're guaranteed to
have the directories created that are specified in configuration.