bug 1737691: fix MinidumpStackwalkRule to create cache directories
This fixes the MinidumpStackwalkRule to create the symbol cache/tmp
directories it needs if they don't exist at startup. This fixes the
problem where the directories created in Dockerfile may not match the
directories specified in configuration.

It also puts the onus on the MinidumpStackwalkRule to create the
directories rather than force the minidump-stackwalk binary to do it.
If there are problems, it'll be easier to see the errors.